fk_ns_second多智能体ver1.1/requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
langchain-core>=0.3,<0.4
langchain-openai>=0.3,<0.4
tavily-python>=0.5,<0.6
mcp>=1.0,<2.0
fastmcp>=2.0,<3.0
openai>=1.0,<2.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0